[react-ui][doc] Améliorations et refacto
- Déplace la documentation des mécanismes dans le site docusaurus
- Depuis RuleaPage, utilise un lien vers le site publi.codes pour la description du mecanisme (plus léger, plus simple, et toujours à jour)
- Supprime le composant Overlay
- Améliore le style du bac à sable

fix #78, fix #12
