+++ packages/react-ui/package.json
@@ -13,19 +13,17 @@
 		}
 	},
 	"scripts": {
-		"build:mecanisms": "yarn run --silent js-yaml ./source/data/mecanisms.yaml >| ./source/data/mecanisms.json",
 		"prepare": "yarn run build:mecanisms",
 		"clean": "rimraf dist node_modules data",
 		"build": "tsup source/index.tsx --format cjs,esm --dts --inject ./react-shim.js --sourcemap",
+		"build:watch": "tsup source/index.tsx --format cjs,esm --watch --inject ./react-shim.js --sourcemap  ",
 		"test": "echo \"Error: no test specified\""
 	},
 	"files": [
 		"dist"
 	],
 	"dependencies": {
-		"focus-trap-react": "^8.8.2",
-		"styled-components": "^5.1.0",
-		"yaml": "^1.9.2"
+		"styled-components": "^5.1.0"
 	},
 	"peerDependencies": {
 		"publicodes": "1.0.0-beta.30",
@@ -34,7 +32,6 @@
 	"devDependencies": {
 		"@types/react": "^17.0.0",
 		"@types/styled-components": "^5.1.10",
-		"js-yaml": "^4.1.0",
 		"tsup": "^5.11.11",
 		"typescript": "^4.5.5"
 	}
