+++ website/package.json
@@ -15,8 +15,9 @@
 		"prepare": "cd .. && yarn install && yarn run prepare"
 	},
 	"dependencies": {
-		"@docusaurus/core": "^2.0.0-beta.9",
-		"@docusaurus/preset-classic": "^2.0.0-beta.9",
+		"@docusaurus/core": "^2.0.0-beta.15",
+		"@docusaurus/preset-classic": "^2.0.0-beta.15",
+		"@docusaurus/remark-plugin-npm2yarn": "^2.0.0-beta.15",
 		"@mdx-js/react": "^1.6.21",
 		"@svgr/webpack": "^5.5.0",
 		"clsx": "^1.1.1",
