--- website/src/pages/index.js
@@ -1,39 +1,0 @@
-import React from 'react'
-import clsx from 'clsx'
-import Layout from '@theme/Layout'
-import Link from '@docusaurus/Link'
-import useDocusaurusContext from '@docusaurus/useDocusaurusContext'
-import styles from './index.module.css'
-import Intro from '../components/intro.mdx'
-
-function HomepageHeader() {
-	const { siteConfig } = useDocusaurusContext()
-	return (
-		<header className={clsx('hero hero--primary', styles.heroBanner)}>
-			<div className="container">
-				<h1 className="hero__title">{siteConfig.title}</h1>
-				<p className="hero__subtitle">{siteConfig.tagline}</p>
-				<div className={styles.buttons}>
-					<Link className="button button--secondary button--lg" to="/docs/se-lancer">
-						Découvrir le langage
-					</Link>
-				</div>
-			</div>
-		</header>
-	)
-}
-
-export default function Home() {
-	const { siteConfig } = useDocusaurusContext()
-	return (
-		<Layout
-			title={`Langage ${siteConfig.title}`}
-			description="Langage et plateforme pour des calculs auto-documentés"
-		>
-			<HomepageHeader />
-			<main className={styles.main}>
-				<Intro />
-			</main>
-		</Layout>
-	)
-}
