+++ website/src/pages/index.module.css
@@ -4,26 +4,6 @@
  * CSS files with the .module.css suffix will be treated as CSS modules
  * and scoped locally.
  */
-
-.heroBanner {
-	padding: 4rem 0;
-	text-align: center;
-	position: relative;
-	overflow: hidden;
-}
-
-@media screen and (max-width: 966px) {
-	.heroBanner {
-		padding: 2rem;
-	}
-}
-
-.buttons {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-}
-
 main {
 	max-width: 850px;
 	margin-top: 4rem;
