+++ website/src/pages/statistiques.tsx
@@ -1,6 +1,5 @@
-import React from 'react'
+import { useColorMode } from '@docusaurus/theme-common'
 import Layout from '@theme/Layout'
-import useThemeContext from '@theme/hooks/useThemeContext'
 
 export default function Statistiques() {
 	return (
@@ -15,7 +14,7 @@ export default function Statistiques() {
 }
 
 function IFrame() {
-	const { isDarkTheme } = useThemeContext()
+	const { isDarkTheme } = useColorMode()
 
 	return (
 		<>
